requirements.txt
aiogram==2.14
openai
requests
python-dotenv
